+++ packages/soft-config/src/puck/actions/useSetDefaultVersion.tsx
@@ -25,7 +25,7 @@ export const useSetDefaultVersion = () => {
 
     setSoftComponentDefaultVersion(componentName, version);
     
-    triggerAction({
+    void triggerAction({
       type: "setDefaultVersion",
       payload: {
         id: componentName,